+++ src/svlog/typeck.rs
@@ -85,6 +85,7 @@ pub(crate) fn type_of<'gcx>(
                 }
             }
             hir::ExprKind::Builtin(hir::BuiltinCall::Clog2(_)) => Ok(cx.mkty_int(32)),
+            hir::ExprKind::Ternary(_cond, true_expr, _false_expr) => cx.type_of(true_expr, env),
             _ => {
                 error!("{:#?}", hir);
                 cx.unimp_msg("type analysis of", &hir)