+++ tests/svlog/const_ops.sv
@@ -16,6 +16,8 @@ module A #(int a = 42, int b = 19);
 	B #(a <<< b) b15();
 	B #(a >> b) b16();
 	B #(a >>> b) b17();
+	B #(a == b ? 42 : 9001) b18();
+	B #(a != b ? 42 : 9001) b19();
 endmodule
 
 module B #(int K);
@@ -91,6 +93,14 @@ endmodule
 //|     %x = sig i32 0
 //| }
 //|
+//| entity @B.param18 () () {
+//|     %x = sig i32 9001
+//| }
+//|
+//| entity @B.param19 () () {
+//|     %x = sig i32 42
+//| }
+//|
 //| entity @A () () {
 //|     %b1 = inst @B.param1 () ()
 //|     %b2 = inst @B.param2 () ()
@@ -109,4 +119,6 @@ endmodule
 //|     %b15 = inst @B.param15 () ()
 //|     %b16 = inst @B.param16 () ()
 //|     %b17 = inst @B.param17 () ()
+//|     %b18 = inst @B.param18 () ()
+//|     %b19 = inst @B.param19 () ()
 //| }