+++ tests/svlog/exprs.sv
@@ -20,6 +20,7 @@ module A;
 		#1ns a = !b;
 		#1ns a = b && c;
 		#1ns a = b || c;
+		#1ns a = b == c ? 42 : 9001;
 		// a = b * c;
 		// a = b / c;
 		// a = b & c;