+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for the `$clog2` builtin function. ([#81](https://github.com/fabianschuiki/moore/issues/81))
 - Support for non-blocking assignments with and without delay. ([#82](https://github.com/fabianschuiki/moore/issues/82))
 - Support for based integer literals, and `'0`, and `'1`. ([#84](https://github.com/fabianschuiki/moore/issues/84))
+- Support for the ternary operator in processes and functions. ([#83](https://github.com/fabianschuiki/moore/issues/83))
 
 ### Fixed
 - Fix blocking assignments, making them actually block. ([#78](https://github.com/fabianschuiki/moore/issues/78))